feat: Add Category and Client many-to-many relationship
This commit adds a many-to-many relationship between the Category and Client entities. It introduces a new property `$Category` in the Client entity and a new property `$client` in the Category entity. The necessary annotations and methods have been added to establish the relationship. This feature allows clients to be associated with multiple categories and categories to have multiple clients.

Closes #456